src/streaming/kafka_producer.py
# =========================
# streaming/kafka_producer.py
# =========================
from kafka import KafkaProducer
import json

producer = KafkaProducer(
    bootstrap_servers="localhost:9092",
    value_serializer=lambda v: json.dumps(v).encode("utf-8")
)


def send_news(news):
    producer.send("financial-news", news)
