pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "financial-sentiment-ai"
version = "1.0.0"
description = "Production-ready financial sentiment analysis with LoRA, RAG, and Kafka"
authors = [
    {name = "Your Name", email = "you@example.com"}
]
dependencies = [
    "transformers>=4.35.0",
    "peft>=0.6.0",
    "torch>=2.1.0",
    "faiss-cpu>=1.8.0",
    "kafka-python>=2.0.2",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "pydantic>=2.5.0"
]

[tool.setuptools.packages.find]
where = ["."]
